requirements.txt
pynacl
psycopg2-binary
scipy
pandas
numpy
python-dotenv
pytest
